Number of changes
Notably:
Created vec.{c,h} to provide generic vector functions.
Vecs store capacity now.  Capacity and length are stored as size_t.
We use vec_ naming convention now for vecs.
Fixed major bug in walk_fpos().
Fixed some memory allocation bugs.
main.c has function actual_main().  main() is in ../main.c.  This
  allows for testing to link in with more ease.

+++ src/parse_args.h
@@ -8,9 +8,11 @@
 #ifndef HEADER_GUARD_PARSE_ARGS_H
 #define HEADER_GUARD_PARSE_ARGS_H
 
+#include <stddef.h>
+
 typedef struct arguments {
     /* NOT null */ const char** files;
-    int num_files;
+    size_t num_files;
 } arguments;
 
 int parse_args(int argc, const char** argv,
